Separate FormCustomizer out from PaymentForm.
