+++ app/src/main/java/checkout/checkout_android/DemoActivity.java
@@ -7,6 +7,7 @@ import android.support.v7.app.AlertDialog;
 import android.util.Log;
 
 import com.checkout.sdk.CheckoutClient;
+import com.checkout.sdk.FormCustomizer;
 import com.checkout.sdk.core.TokenResult;
 import com.checkout.sdk.paymentform.PaymentForm;
 import com.checkout.sdk.utils.CardUtils.Cards;
@@ -24,7 +25,7 @@ public class DemoActivity extends Activity {
                 mPaymentForm.clearForm(); // clear the form
                 String id = ((TokenResult.TokenResultSuccess) tokenResult).getResponse().getId();
                 displayMessage("Token", id);
-                Log.e("JOHN", "Token: " + id);
+                Log.e("TOKEN", "Token: " + id);
             } else if (tokenResult instanceof TokenResult.TokenResultTokenisationFail) {
                 String errorCode = ((TokenResult.TokenResultTokenisationFail) tokenResult).getError().getErrorCode();
                 displayMessage("Token Error", errorCode);
@@ -42,27 +43,16 @@ public class DemoActivity extends Activity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_demo);
 
-        mPaymentForm = findViewById(R.id.checkout_card_form);
         CheckoutClient checkoutClient = new CheckoutClient(
                 this,
                 "pk_test_6e40a700-d563-43cd-89d0-f9bb17d35e73",
                 Environment.SANDBOX,
                 callback);
-        mPaymentForm
-                .setAcceptedCard(new Cards[]{Cards.VISA, Cards.MASTERCARD})
-                .initialize(checkoutClient);
 
-        // TODO: Make it look more like:
-        // CheckoutClient checkoutClient = new CheckoutClient(this,
-//        "pk_test_6e40a700-d563-43cd-89d0-f9bb17d35e73",
-//                tokenListener,
-//                Environment.SANDBOX
-//        );
-//        UiCustomizer uiCustomizer = new UiCustomizer.Builder()
-//                .acceptedCards(Arrays.asList(Cards.VISA, Cards.MASTERCARD))
-//                .build();
-//        mPaymentForm = findViewById(R.id.checkout_card_form);
-//        mPaymentForm.initialize(checkoutClient, mFormListener, uiCustomizer);
+        FormCustomizer formCustomizer = new FormCustomizer()
+                .setAcceptedCard(new Cards[]{Cards.VISA, Cards.MASTERCARD});
+        mPaymentForm = findViewById(R.id.checkout_card_form);
+        mPaymentForm.initialize(checkoutClient, formCustomizer);
     }
 
     private void displayMessage(String title, String message) {
